requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
